Clean-up the Makefile, add a posix strdup fallback
Removes the extra debug build target. The build configuration
is dependent on the DEBUG flag, which can be enabled using `make DEBUG=1`.
Adds guards to util.h/util.c to provide the strdup/strndup implementation
for the specific platform. _POSIX_C_SOURCE will be defined when you
are not using a gcc compiler. Also renames index to index0 in util.c
to prevent the mix-up with the index function.