Update import system
Already imported modules/files are no longer reimported.
During an import, the compiler checks, if the file has already been
imported. If not, the import name is inserted into an import map
and the file will be compiled.
The AST of the imported file will not be generated during the parsing
phase as before.