Make the parser independent of the lexer, overall update
Renames AST_TOPLEVEL to AST_BLOCK.
Updates the parser to copy and keep the data of the lexer.
The lexer can now be immediately freed after the creation of the tokens.
This removes the need for subparsers in the compiler to keep the tokens alive.
Only one parser is in use. Updates the standard library and the ast_free function
to work correctly with the new setting. Also removes dead code and the
bytecode_buffer_free function to handle NULL input.